move hint types to types.ts
