--- src/background/hint/HintTarget.ts
@@ -1,7 +1,0 @@
-type HintTarget = {
-  frameId: number;
-  element: string;
-  tag: string;
-};
-
-export default HintTarget;
