+++ src/background/hint/OpenSourceHintAction.ts
@@ -1,7 +1,6 @@
 import { injectable, inject } from "inversify";
 import type HintClient from "../clients/HintClient";
-import type HintTarget from "./HintTarget";
-import type HintAction from "./HintAction";
+import type { HintTarget, HintAction } from "./types";
 import type TabPresenter from "../presenters/TabPresenter";
 
 @injectable()
