+++ src/background/usecases/HintKeyUseCase.ts
@@ -1,6 +1,6 @@
 import { injectable, inject } from "inversify";
 import type HintClient from "../clients/HintClient";
-import type HintTarget from "../hint/HintTarget";
+import type { HintTarget } from "../hint/types";
 import type HintRepository from "../repositories/HintRepository";
 import type HintActionFactory from "../hint/HintActionFactory";
